+++ database/migrations/2023_04_27_150017_add_field_in_table_pmieducar_matricula_turma.php
@@ -8,15 +8,19 @@ return new class() extends Migration
 {
     public function up(): void
     {
-        Schema::table('pmieducar.matricula_turma', function (Blueprint $table) {
-            $table->integer('cod_curso_profissional')->nullable();
-        });
+        if (!Schema::hasColumn('pmieducar.matricula_turma', 'cod_curso_profissional')) {
+            Schema::table('pmieducar.matricula_turma', function (Blueprint $table) {
+                $table->integer('cod_curso_profissional')->nullable();
+            });
+        }
     }
 
     public function down(): void
     {
-        Schema::table('pmieducar.matricula_turma', function (Blueprint $table) {
-            $table->dropColumn('cod_curso_profissional');
-        });
+        if (Schema::hasColumn('pmieducar.matricula_turma', 'cod_curso_profissional')) {
+            Schema::table('pmieducar.matricula_turma', function (Blueprint $table) {
+                $table->dropColumn('cod_curso_profissional');
+            });
+        }
     }
 };
