+++ database/migrations/2023_05_08_221446_add_field_in_table_pmieducar_modules_professor_turma.php
@@ -8,15 +8,19 @@ return new class() extends Migration
 {
     public function up(): void
     {
-        Schema::table('modules.professor_turma', function (Blueprint $table) {
-            $table->smallInteger('outras_unidades_curriculares_obrigatorias')->default(0);
-        });
+        if (!Schema::hasColumn('modules.professor_turma', 'outras_unidades_curriculares_obrigatorias')) {
+            Schema::table('modules.professor_turma', function (Blueprint $table) {
+                $table->smallInteger('outras_unidades_curriculares_obrigatorias')->default(0);
+            });
+        }
     }
 
     public function down(): void
     {
-        Schema::table('modules.professor_turma', function (Blueprint $table) {
-            $table->dropColumn('outras_unidades_curriculares_obrigatorias')->default(0);
-        });
+        if (Schema::hasColumn('modules.professor_turma', 'outras_unidades_curriculares_obrigatorias')) {
+            Schema::table('modules.professor_turma', function (Blueprint $table) {
+                $table->dropColumn('outras_unidades_curriculares_obrigatorias')->default(0);
+            });
+        }
     }
 };
